requirements.txt
praw==3.5.0
pillow==3.4.2
numpy==1.11.2
matplotlib==1.5.3
wordcloud==1.2.1
imgurpython==1.1.7
nltk==3.2.1
PyYAML==3.12